Reorganize code for Coral sketches
