+++ Coral/Maze/index.html
@@ -6,14 +6,14 @@
     <script
       language="javascript"
       type="text/javascript"
-      src="../libraries/p5.min.js"
+      src="../../libraries/p5.min.js"
     ></script>
     <script type="module">
       import { sketch } from "./MazeCoral.js";
-      import "../common/sketchbook-navbar.js";
+      import "../../common/sketchbook-navbar.js";
       const p5_sketch = new p5(sketch, "sketch");
     </script>
-    <link rel="stylesheet" href="../styles/layout.css" />
+    <link rel="stylesheet" href="../../styles/layout.css" />
     <style>
       img {
         display: block;
@@ -33,7 +33,7 @@
       <div class="controls">
         <label for="import"
           >Import coral tileset from
-          <a href="../Coral/">Coral Tile Editor</a></label
+          <a href="../TileEditor/">Coral Tile Editor</a></label
         ><br />
         <input
           type="file"
