+++ Coral/TileEditor/index.html
@@ -6,14 +6,14 @@
     <script
       language="javascript"
       type="text/javascript"
-      src="../libraries/p5.min.js"
+      src="../../libraries/p5.min.js"
     ></script>
     <script type="module">
-      import { sketch } from "./Coral.js";
-      import "../common/sketchbook-navbar.js";
+      import { sketch } from "./TileEditor.js";
+      import "../../common/sketchbook-navbar.js";
       const p5_sketch = new p5(sketch, "sketch");
     </script>
-    <link rel="stylesheet" href="../styles/layout.css" />
+    <link rel="stylesheet" href="../../styles/layout.css" />
     <style>
       img {
         display: block;
@@ -24,7 +24,7 @@
   <body>
     <sketchbook-navbar></sketchbook-navbar>
     <div class="container">
-      <h1>WIP Coral Tile Editor (2025)</h1>
+      <h1>Coral Tile Editor (2025)</h1>
       <div id="sketch"></div>
       <div class="controls"><button id="export">Export JSON</button></div>
       <div class="description">
